move consumer middleware inside consumer dir
