+++ consumer/middleware/expire_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/makasim/amqpextra"
-	"github.com/makasim/amqpextra/middleware"
+	"github.com/makasim/amqpextra/consumer/middleware"
 	"github.com/streadway/amqp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
