+++ consumer/middleware/has_correlation_id_test.go
@@ -6,7 +6,7 @@ import (
 
 	"fmt"
 
-	"github.com/makasim/amqpextra/middleware"
+	"github.com/makasim/amqpextra/consumer/middleware"
 	"github.com/streadway/amqp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
