+++ consumer/middleware/has_reply_to_test.go
@@ -6,7 +6,7 @@ import (
 
 	"fmt"
 
-	"github.com/makasim/amqpextra/middleware"
+	"github.com/makasim/amqpextra/consumer/middleware"
 	"github.com/streadway/amqp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
