+++ consumer/middleware/logger.go
