+++ consumer/middleware/recover.go
