+++ consumer/middleware/recover_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/makasim/amqpextra"
-	"github.com/makasim/amqpextra/middleware"
+	"github.com/makasim/amqpextra/consumer/middleware"
 	"github.com/streadway/amqp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
